refactor: ShoppingCard 스타일 분리
